fix(web): break offline import cycle and resolve dead-code gate
C1: extracted replayIntent (api-importing fn) from intents.ts into new
replay.ts, breaking the api→foldPending→intents→api cycle. drainer.ts
and drainer.test.ts updated to import/mock from ./replay.

C2: wired useOnlineStatus into SyncStatusBadge so the hook is consumed
by production code; badge now renders an Offline indicator when offline.
Test updated with mock for useOnlineStatus and an offline test case.

C3: added fake-indexeddb to ignoreDependencies in .fallowrc.json
(used only via fake-indexeddb/auto in test files, which fallow ignores).
