+++ packages/web/src/offline/drainer.test.ts
@@ -2,7 +2,7 @@ import 'fake-indexeddb/auto';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 const { replayIntent } = vi.hoisted(() => ({ replayIntent: vi.fn() }));
-vi.mock('./intents', () => ({ replayIntent }));
+vi.mock('./replay', () => ({ replayIntent }));
 
 import { drainOutbox } from './drainer';
 import { outboxStore } from './outboxStore';
